+++ android-sdk-ui/src/main/java/com/braze/ui/inappmessage/InAppMessageOperation.kt
@@ -4,7 +4,7 @@ import androidx.annotation.Keep
 import java.util.*
 
 enum class InAppMessageOperation {
-    DISPLAY_NOW, DISPLAY_LATER, DISCARD;
+    DISPLAY_NOW, DISPLAY_LATER, DISCARD, REENQUEUE;
 
     @Keep
     companion object {